2021-03-16 WinForm Arduino Sensor Monitoring

+++ WinformApp/WinFormAdvancedBank/AddressInfoApp/FrmMain.cs
@@ -100,22 +100,25 @@ namespace AddressInfoApp
                 MessageBox.Show("데이터를 선택하세요.");
                 return;
             }
-            using (SqlConnection conn = new SqlConnection(connString))
-            {
-                if (conn.State == ConnectionState.Closed) conn.Open();
-                string query = $"DELETE FROM AddressTbl WHERE idx = {result}";
-                SqlCommand cmd = new SqlCommand(query, conn);
-                if (cmd.ExecuteNonQuery() == 1)
-                {
-                    MessageBox.Show("삭제 성공!");
-                }
-                else MessageBox.Show("삭제 실패!");
-            }
 
+            if (MessageBox.Show("삭제하시겠습니까?", "삭제", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            {
+                using (SqlConnection conn = new SqlConnection(connString))
+                {
+                    if (conn.State == ConnectionState.Closed) conn.Open();
+                    string query = $"DELETE FROM AddressTbl WHERE idx = {result}";
+                    SqlCommand cmd = new SqlCommand(query, conn);
+                    if (cmd.ExecuteNonQuery() == 1)
+                    {
+                        MessageBox.Show("삭제 성공!");
+                    }
+                    else MessageBox.Show("삭제 실패!");
+                }
+            }
             RefreshData();
             ClearInput();
         }
-        // 삭제 버튼에 질문 추가 
+        
         private void btnSearch_Click(object sender, EventArgs e)
         {
